Convert VcdFile into GObject
