+++ src/lx2.c
@@ -23,6 +23,7 @@
 #include "busy.h"
 #include "hierpack.h"
 #include "fst.h"
+#include "gw-vcd-file.h"
 
 /*
  * actually import an lx2 trace but don't do it if it's already been imported
@@ -31,7 +32,7 @@ void import_lx2_trace(GwNode *np)
 {
     switch (GLOBALS->is_lx2) {
         case LXT2_IS_VLIST:
-            import_vcd_trace(GLOBALS->vcd_file, np);
+            gw_vcd_file_import_trace(GW_VCD_FILE(GLOBALS->dump_file), np);
             return;
         case LXT2_IS_FST:
             import_fst_trace(GLOBALS->fst_file, np);
@@ -52,7 +53,7 @@ void lx2_set_fac_process_mask(GwNode *np)
 {
     switch (GLOBALS->is_lx2) {
         case LXT2_IS_VLIST:
-            vcd_set_fac_process_mask(GLOBALS->vcd_file, np);
+            gw_vcd_file_set_fac_process_mask(GW_VCD_FILE(GLOBALS->dump_file), np);
             return;
         case LXT2_IS_FST:
             fst_set_fac_process_mask(GLOBALS->fst_file, np);
@@ -70,7 +71,7 @@ void lx2_import_masked(void)
 {
     switch (GLOBALS->is_lx2) {
         case LXT2_IS_VLIST:
-            vcd_import_masked(GLOBALS->vcd_file);
+            gw_vcd_file_import_masked(GW_VCD_FILE(GLOBALS->dump_file));
             return;
         case LXT2_IS_FST:
             fst_import_masked(GLOBALS->fst_file);
