+++ src/meson.build
@@ -52,6 +52,7 @@ gtkwave_sources = [
     'gw-mouseover.c',
     'gw-named-marker-dialog.c',
     'gw-time-display.c',
+    'gw-vcd-file.c',
     'gw-wave-view-traces.c',
     'gw-wave-view.c',
     'hierpack.c',
